chore: imports cleanup, extract helper functions
