+++ admin/src/tables/SerpContentGapTable.jsx
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useMemo, memo } from 'react';
 import { __ } from '@wordpress/i18n';
-import { queryTypes, queryStatuses, queryScheduleIntervals, queryHeaders, queryLevels, queryIntents } from '../lib/serpQueryColumns';
 
 import {
 	useInfiniteFetch,
@@ -16,128 +15,26 @@ import {
 
 import useTableStore from '../hooks/useTableStore';
 import useTablePanels from '../hooks/useTablePanels';
-import DescriptionBox from '../elements/DescriptionBox';
 import useChangeRow from '../hooks/useChangeRow';
+import useSerpGapCompare from '../hooks/useSerpGapCompare';
+import { countriesList, countriesListForSelect } from '../api/fetchCountries';
 
-import hexToHSL from '../lib/hexToHSL';
+import { getTooltipUrlsList } from '../lib/elementsHelpers';
+import { colorRankingBackground, colorRankingInnerStyles, emptyUrls, preprocessUrls } from '../lib/serpContentGapHelpers';
+import { queryTypes, queryStatuses, queryScheduleIntervals, queryHeaders, queryLevels, queryIntents } from '../lib/serpQueryColumns';
+
+import Box from '@mui/joy/Box';
+import Tooltip from '@mui/joy/Tooltip';
+import Button from '@mui/joy/Button';
+
+import DescriptionBox from '../elements/DescriptionBox';
 import GapDetailPanel from '../components/detailsPanel/GapDetailPanel';
-import { postFetch } from '../api/fetching';
-import { setNotification } from '../hooks/useNotifications';
 
 import '../assets/styles/layouts/ContentGapTableCells.scss';
-import { Box, Tooltip } from '@mui/joy';
-import { countriesList, countriesListForSelect } from '../api/fetchCountries';
-import { getTooltipUrlsList } from '../lib/elementsHelpers';
-import Button from '@mui/joy/Button';
-import useSerpGapCompare from '../hooks/useSerpGapCompare';
 
 const paginationId = 'query_id';
 const optionalSelector = '';
 const defaultSorting = [ { key: 'comp_intersections', dir: 'DESC', op: '<' } ];
-
-const maxProcessingAttempts = 3;
-
-const validateResults = ( results ) => {
-	const withError = Object.values( results ).filter( ( data ) => data.status === 'error' );
-	if ( withError.length ) {
-		setNotification( 'serp-gap/prepare/download-failed', {
-			title: __( 'Download of some URLs failed.' ),
-			message: __( 'Some data will not be present in table.' ),
-			status: 'error',
-		} );
-	}
-	return results;
-};
-
-const preprocessUrls = async ( { urls, parse_headers }, processing = 0 ) => {
-	try {
-		const response = await postFetch( 'serp-gap/prepare', { urls, parse_headers }, { skipErrorHandling: true } );
-
-		if ( response.status === 200 ) {
-			const results = await response.json();
-			return validateResults( results );
-		}
-
-		if ( response.status === 404 ) {
-			if ( processing < maxProcessingAttempts ) {
-				return await preprocessUrls( { urls, parse_headers }, processing + 1 );
-			}
-			const results = await response.json();
-			return validateResults( results );
-		}
-
-		if ( response.status === 400 ) {
-			const results = await response.json();
-			return validateResults( results );
-		}
-
-		throw new Error( 'Failed to process URLs.' );
-	} catch ( error ) {
-		setNotification( 'serp-gap/prepare/error', { message: error.message, status: 'error' } );
-		return [];
-	}
-};
-
-const colorRankingBackground = ( val ) => {
-	const value = Number( val );
-	const okColor = '#EEFFEE'; // light green
-	const failColor = '#FFEEEE'; // light red
-
-	if ( typeof value === 'number' && value < 0 ) {
-		return { backgroundColor: '#EEEEEE' };
-	}
-
-	if ( typeof value !== 'number' || value === 0 || value > 50 ) {
-		const { h, s } = hexToHSL( failColor );
-		const l = 70;
-		return { backgroundColor: `hsl(${ h }, ${ s }%, ${ l }%)` };
-	}
-
-	if ( value > 0 && value <= 10 ) {
-		const { h, s } = hexToHSL( okColor );
-		const l = ( 70 + ( value * 2 ) );
-		return { backgroundColor: `hsl(${ h }, ${ s }%, ${ l }%)` };
-	}
-	const { h, s } = hexToHSL( failColor );
-	const l = ( 100 - ( value / 3 ) );
-	return { backgroundColor: `hsl(${ h }, ${ s }%, ${ l }%)` };
-};
-
-const colorRankingInnerStyles = ( { position, words } ) => {
-	const styles = {};
-
-	if ( position !== null ) {
-		const value = Number( position );
-		if ( value > 0 && value <= 50 ) {
-			styles[ '--position-color' ] = '#000000';
-		} else {
-			styles[ '--position-color' ] = '#FFFFFF';
-		}
-	}
-
-	if ( words !== null ) {
-		const value = Number( words );
-		if ( value > 0 && value <= 20 ) {
-			styles[ '--words-backgroundColor' ] = '#2CA34D'; // green light
-		} else if ( value > 20 && value <= 50 ) {
-			styles[ '--words-backgroundColor' ] = '#187E36'; // green medium
-		} else if ( value > 50 ) {
-			styles[ '--words-backgroundColor' ] = '#106228'; // green dark
-		}
-	}
-
-	return styles;
-};
-
-const emptyUrls = ( urls ) => {
-	if ( ! urls || ( urls && Object.keys( urls ).length === 0 ) ) {
-		return true;
-	}
-
-	return Object.values( urls ).filter( ( url ) => url !== '' ).length
-		? false
-		: true;
-};
 
 const SerpContentGapTable = memo( ( { slug } ) => {
 	const setFetchOptions = useTablePanels( ( state ) => state.setFetchOptions );
@@ -502,7 +399,7 @@ const TableContent = memo( ( { slug } ) => {
 		) );
 
 		return { header, columns };
-	}, [ updateFetchOptions, slug, urls, columnHelper, __ ] );
+	}, [ columnHelper, urls, updateFetchOptions, updateRow, compareUrls, slug ] );
 
 	const { columns } = columnsDef;
